+++ src/sensors.cpp
@@ -6,10 +6,10 @@ void ADCsensor::init(mcp3008Spi* adc, int channel)
 	channel_ = channel;
 }
 
-double ADCsensor::getValue()
+unsigned int ADCsensor::getValue()
 {
 	unsigned char data[3];
-	int adcVal;
+	unsigned int adcVal;
 	
 	data[0] = 1;  //  first byte transmitted -> start bit
 	data[1] = 0b10000000 |( ((channel_ & 7) << 4)); // second byte transmitted -> (SGL/DIF = 1, D2=D1=D0=0)
